refactor: make command destructurization
