+++ command/exit.js
@@ -1,0 +1,5 @@
+import process from "node:process";
+
+export const exit = () => {
+    process.exit(0);
+};
